Mooving to utils getToken function
